migrate device item
