+++ src/HW/LedgerConnect/DeviceItem/index.ts
@@ -1,3 +1,1 @@
-// @flow
-
 export * from './DeviceItem'
