+++ storybook/storyLoader.js
@@ -5,7 +5,6 @@
 
 function loadStories() {
 	require('../legacy/components/Delegation/FlawedWalletScreen.stories');
-	require('../legacy/components/Ledger/DeviceItem/DeviceItem.stories');
 	require('../legacy/components/MaintenanceScreen.stories');
 	require('../legacy/components/UiKit/Button.stories');
 	require('../legacy/components/UiKit/Checkbox.stories');
@@ -22,6 +21,7 @@ function loadStories() {
 	require('../src/FirstRun/LanguagePickerScreen/LanguagePickerScreen.stories');
 	require('../src/FirstRun/TermsOfServiceScreen/TermsOfServiceScreen.stories');
 	require('../src/HW/Instructions/Instructions.stories');
+	require('../src/HW/LedgerConnect/DeviceItem/DeviceItem.stories');
 	require('../src/HW/LedgerConnect/LedgerConnect.stories');
 	require('../src/HW/LedgerTransportSwitchModal/LedgerTransportSwitchModal.stories');
 	require('../src/Login/CustomPinLoginScreen.stories');
@@ -86,7 +86,6 @@ function loadStories() {
 
 const stories = [
 	'../legacy/components/Delegation/FlawedWalletScreen.stories',
-	'../legacy/components/Ledger/DeviceItem/DeviceItem.stories',
 	'../legacy/components/MaintenanceScreen.stories',
 	'../legacy/components/UiKit/Button.stories',
 	'../legacy/components/UiKit/Checkbox.stories',
@@ -103,6 +102,7 @@ const stories = [
 	'../src/FirstRun/LanguagePickerScreen/LanguagePickerScreen.stories',
 	'../src/FirstRun/TermsOfServiceScreen/TermsOfServiceScreen.stories',
 	'../src/HW/Instructions/Instructions.stories',
+	'../src/HW/LedgerConnect/DeviceItem/DeviceItem.stories',
 	'../src/HW/LedgerConnect/LedgerConnect.stories',
 	'../src/HW/LedgerTransportSwitchModal/LedgerTransportSwitchModal.stories',
 	'../src/Login/CustomPinLoginScreen.stories',
